Move edit logic to child components - mw
